Made some requested changes
